Use TCP keep-alive probing on the event stream connection
If you are connecting to the event stream through some load balancing device,
there is a good chance your connection will get dropped after some time if no
traffic is crossing the wire. In the case of an AWS ELB, this interval is 60
seconds. Enable TCP keepalive and set the idle time and time between probes to
50 seconds to prevent connection dropping.